Refactor ErrorTab
- Break into container/component
- Add (bootstrap) WrapperContainer component to handle repeated divs.
- Have ErrorTab handle 404 errors.
- Auto popup rLogin if it is not connected.
