+++ src/app/tabs/index.js
@@ -1,7 +1,7 @@
 import resolve from './resolve';
 import search from './search';
 import registrar from './registrar';
-import error from './error';
+import error from './error/ErrorTabComponent';
 import newAdmin from './newAdmin';
 import faqTab from './faq';
 
@@ -21,5 +21,5 @@ export { ResolveTab } from './resolve';
 export { SearchTab } from './search';
 export { RegistrarTab } from './registrar';
 export { default as SetUpTab } from './SetUpTab';
-export { default as ErrorTab } from './error';
+export { default as ErrorTab } from './error/ErrorTabContainer';
 export { default as FaqTab } from './faq';
